Add macros, update includes, update CMake, prepare for Kokkos.
Added macros.h which contains an Assert macro and will also contain macros for working with Kokkos in the future.

Updated includes so that every file includes macros.h, and macros.h will include relevant <vector>, <iostream>, Kokkos includes, etc.

Updated the CMake process to be more customizable. It may now support Debug or Release mode and check for Kokkos.

Started preparing for an optional Kokkos dependency. The goal is to be able to run some fractal computations on the GPU.